Preserve cache timestamps across cycle wrap

+++ src/gpgpu-sim/l2cache.h
@@ -201,7 +201,7 @@ class memory_sub_partition {
   void get_L2cache_sub_stats_pw(struct cache_sub_stats_pw &css) const;
   void clear_L2cache_stats_pw();
 
-  void force_l2_tag_update(new_addr_type addr, unsigned time,
+  void force_l2_tag_update(new_addr_type addr, unsigned long long time,
                            mem_access_sector_mask_t mask) {
     m_L2cache->force_tag_access(addr, m_memcpy_cycle_offset + time, mask);
     m_memcpy_cycle_offset += 1;
